pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "py-cord[voice]>=2.6.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
